requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Science & Arrays
numpy<2.0.0
pandas>=1.5.0
scipy>=1.9.0
scikit-learn>=1.2.0

# Computer Vision & Image Processing
opencv-python-headless>=4.7.0
Pillow>=9.4.0
scikit-image>=0.20.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging, Config & Utilities
tqdm>=4.65.0
PyYAML>=6.0
tensorboard>=2.12.0
wandb>=0.15.0
omegaconf>=2.3.0
black>=23.3.0

# Advanced Models & Transfer Learning (Classification, Transformers)
timm>=0.9.0
transformers>=4.30.0
datasets>=2.13.0
huggingface-hub>=0.16.0
accelerate>=0.21.0
peft>=0.4.0

# Testing & Profiling
pytest>=7.0.0

# Deployment
onnxruntime>=1.16.0

# Optional packages for specific cell pathology tasks (Uncomment to install)
# monai              # Deep learning for medical imaging
ultralytics>=8.0.0   # For YOLO series detection and segmentation
# segment-anything   # For Meta's SAM (Segment Anything Model)
